requirements.txt
streamlit
pandas
openpyxl
Pillow
reportlab
plotly
altair
numpy